+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming Release
 * Streams
   * Internet radio can now play playlists from the `pls`, `m3u8`, and `m3u` formats.
+  * Fixed bug where internet radio process would not completely stop when the stream was stopped.
 
 ## 0.4.1
 * Web App